katy/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
validators==0.18.2
weaviate-client==3.4.2
